+++ apps/backend/fixtures/lambda/daily-alert.json
@@ -1,0 +1,1 @@
+{}
