+++ apps/backend/src/database/datasource.ts
@@ -3,16 +3,16 @@ import { dbLogger, logLevels } from './helpers';
 import { CashDepositEntity } from '../deposits/entities/cash-deposit.entity';
 import { POSDepositEntity } from '../deposits/entities/pos-deposit.entity';
 import { LocationEntity } from '../location/entities';
-import { FileIngestionRulesEntity } from '../parse/entities/file-ingestion-rules.entity';
 import { FileUploadedEntity } from '../parse/entities/file-uploaded.entity';
-import { ProgramDailyUploadEntity } from '../parse/entities/program-daily-upload.entity';
-import { ProgramRequiredFileEntity } from '../parse/entities/program-required-file.entity';
 import { AlertDestinationEntity } from '../notification/entities/alert-destination.entity';
 import {
   TransactionEntity,
   PaymentEntity,
   PaymentMethodEntity,
 } from '../transaction/entities/index';
+import { FileIngestionRulesEntity } from '../notification/entities/file-ingestion-rules.entity';
+import { ProgramDailyUploadEntity } from '../notification/entities/program-daily-upload.entity';
+import { ProgramRequiredFileEntity } from '../notification/entities/program-required-file.entity';
 
 export default new DataSource({
   type: 'postgres',
