+++ apps/backend/src/parse/parse.controller.ts
@@ -21,6 +21,7 @@ import { PosDepositService } from '../deposits/pos-deposit.service';
 import { AppLogger } from '../logger/logger.service';
 import { TransactionEntity } from '../transaction/entities';
 import { TransactionService } from '../transaction/transaction.service';
+import { NotificationService } from '../notification/notification.service';
 
 @Controller('parse')
 @ApiTags('Parser API')
@@ -29,6 +30,8 @@ export class ParseController {
   constructor(
     @Inject(ParseService)
     private readonly parseService: ParseService,
+    @Inject(NotificationService)
+    private readonly alertService: NotificationService,
     @Inject(TransactionService)
     private readonly transactionService: TransactionService,
     @Inject(CashDepositService)
@@ -111,135 +114,14 @@ export class ParseController {
     const { fileName, fileType, program } = body;
     this.appLogger.log(`Parsing ${fileName} - ${fileType}`);
     const contents = file.buffer.toString();
-
-    const allFiles = await this.parseService.getAllFiles();
-    const allFilenames = new Set(allFiles.map((f) => f.sourceFileName));
-    if (allFilenames.has(fileName)) {
-      throw new BadRequestException({
-        message: 'Invalid filename, this already exists',
-      });
+    //TODO call parse service to process file event    
+    return {
+      fileName,
+      fileType,
+      program,
+      contents,
     }
-
-    // Throws an error if no rules exist for the specified program
-    const rules = await this.parseService.getRulesForProgram(program);
-    if (!rules) {
-      throw new HttpException(
-        `No rules established for program ${program}`,
-        HttpStatus.FORBIDDEN
-      );
     }
-
-    // Creates a new daily status for the rule, if none exist, so that files can be tracked
-    let daily = await this.parseService.getDailyForRule(rules, new Date());
-    if (!daily) {
-      daily = await this.parseService.createNewDaily(rules, new Date());
-    }
-
-    try {
-      // FileType is based on the filename (from Parser) or from the endpoint body
-      if (fileType === FileTypes.SBC_SALES) {
-        this.appLogger.log('Parse and store SBC Sales in DB...', fileName);
-        const garmsSales: TransactionEntity[] =
-          await this.parseService.parseGarmsFile(contents, fileName); // validating step
-        const fileToSave = await this.parseService.saveFileUploaded({
-          sourceFileType: fileType,
-          sourceFileName: fileName,
-          sourceFileLength: garmsSales.length,
-          dailyUpload: daily,
-        });
-        this.appLogger.log(`Transaction count: ${garmsSales.length}`);
-        await this.transactionService.saveTransactions(
-          garmsSales.map((sale) => ({
-            ...sale,
-            fileUploadedEntity: fileToSave,
-          }))
-        );
-      }
-
-      if (fileType === FileTypes.TDI17) {
-        this.appLogger.log('Parse and store TDI17 in DB...', fileName);
-        const cashDeposits = await this.parseService.parseTDICashFile(
-          fileName,
-          program,
-          file.buffer
-        ); // validating step
-        const fileToSave = await this.parseService.saveFileUploaded({
-          sourceFileType: fileType,
-          sourceFileName: fileName,
-          sourceFileLength: cashDeposits.length,
-          dailyUpload: daily,
-        });
-        this.appLogger.log(`Cash Deposits count: ${cashDeposits.length}`);
-        await this.cashDepositService.saveCashDepositEntities(
-          cashDeposits.map((deposit) => ({
-            ...deposit,
-            fileUploadedEntity: fileToSave,
-          }))
-        );
-      }
-
-      if (fileType === FileTypes.TDI34) {
-        this.appLogger.log('Parse and store TDI34 in DB...', fileName);
-        const posEntities = await this.parseService.parseTDICardsFile(
-          fileName,
-          program,
-          file.buffer
-        );
-        const fileToSave = await this.parseService.saveFileUploaded({
-          sourceFileType: fileType,
-          sourceFileName: fileName,
-          sourceFileLength: posEntities.length,
-          dailyUpload: daily,
-        });
-        this.appLogger.log(`POS Deposits count: ${posEntities.length}`);
-        await this.posDepositService.savePOSDepositEntities(
-          posEntities.map((deposit) => ({
-            ...deposit,
-            fileUploadedEntity: fileToSave,
-            timestamp: deposit.timestamp,
-          }))
-        );
-      }
-    } catch (err: unknown) {
-      const message =
-        err instanceof Error
-          ? err.message
-          : `Error with processing ${fileName}`;
-      this.appLogger.log(message);
-      throw new BadRequestException({ message });
-    }
-  }
-
-  /**
-   * Commence daily upload for a specific date for each program area we have in the rules
-   * Called at the start of a parser lambda run
-   */
-  @ApiBody({
-    schema: {
-      type: 'object',
-      properties: {
-        date: {
-          type: 'string',
-          nullable: false,
-          example: '2023-01-01',
-        },
-      },
-    },
-  })
-  @Post('daily-upload')
-  async commenceDailyUpload(@Body() body: { date: string }): Promise<void> {
-    const rules = await this.parseService.getAllRules();
-    for (const rule of rules) {
-      const daily = await this.parseService.getDailyForRule(
-        rule,
-        new Date(body.date)
-      );
-      if (!daily) {
-        await this.parseService.createNewDaily(rule, new Date(body.date));
-      }
-    }
-  }
-
   /**
    * Makes decisions on whether to send an alert (or error log) for our programs for a date
    * Based on the daily program status, and which files are required by the rules
@@ -257,62 +139,9 @@ export class ParseController {
       },
     },
   })
+
   @Post('daily-upload/alert')
   async dailyUploadAlert(@Body() body: { date: Date }): Promise<DailyAlertRO> {
-    const rules = await this.parseService.getAllRules();
-    const dailyAlertPrograms = [];
-    for (const rule of rules) {
-      let daily = await this.parseService.getDailyForRule(
-        rule,
-        new Date(body.date)
-      );
-      if (!daily) {
-        daily = await this.parseService.createNewDaily(
-          rule,
-          new Date(body.date)
-        );
-      }
-      if (daily.success) {
-        dailyAlertPrograms.push({
-          program: rule.program,
-          success: true,
-          alerted: false,
-          missingFiles: [],
-        });
-        continue;
-      }
-      const successStatus = this.parseService.determineDailySuccess(
-        rule,
-        daily.files
-      );
-      if (successStatus.success === true) {
-        await this.parseService.saveDaily({
-          ...daily,
-          success: true,
-        });
-        dailyAlertPrograms.push({
-          program: rule.program,
-          success: true,
-          alerted: false,
-          missingFiles: [],
-        });
-      } else {
-        let alerted = false;
-        if (daily.retries >= rule.retries) {
-          alerted = true;
-        }
-        await this.parseService.saveDaily({
-          ...daily,
-          retries: daily.retries + 1,
-        });
-        dailyAlertPrograms.push({
-          program: rule.program,
-          success: false,
-          alerted,
-          missingFiles: successStatus.missingFiles,
-        });
-      }
-    }
-    return { dailyAlertPrograms, date: body.date };
+    return await this.alertService.dailyUploadAlert(body.date);
   }
 }
